# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "server",
    "description": "",
    "license": {
      "name": "AGPL-3.0-only",
      "identifier": "AGPL-3.0-only"
    },
    "version": "0.0.1"
  },
  "paths": {
...
    "/api/artists/{id}/images/{image_id}": {
      "delete": {
        "tags": [
          "artists"
        ],
        "operationId": "delete_artist_image",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Artist ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "image_id",
            "in": "path",
            "description": "Image ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
...
    "/api/songs/{id}": {
      "get": {
        "tags": [
          "songs"
        ],
        "operationId": "get_song",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Song ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Song detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SongResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "songs"
        ],
        "operationId": "update_song",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Song ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSongRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Updated song",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SongResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Title already taken",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      },
      "delete": {
        "tags": [
          "songs"
        ],
        "operationId": "delete_song",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Song ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
    "/api/songs/{id}/images": {
      "post": {
        "tags": [
          "songs"
        ],
        "operationId": "upload_song_image",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Song ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/ImageUpload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Image uploaded and linked",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SongImageInfo"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Song not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
    "/api/songs/{id}/images/{image_id}": {
      "delete": {
        "tags": [
          "songs"
        ],
        "operationId": "delete_song_image",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Song ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "image_id",
            "in": "path",
            "description": "Image ID",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "session": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ArtistInfo": {
        "type": "object",
        "description": "A minimal artist record embedded in song and performance responses.",
        "required": [
          "id",
          "name"
        ],
        "properties": {
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          }
        }
      },
...
      "ErrorResponse": {
        "type": "object",
        "description": "Body returned for all error responses.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
...
      "ImageUpload": {
        "type": "object",
        "description": "Placeholder schema for image multipart upload bodies.",
        "required": [
          "file",
          "kind"
        ],
        "properties": {
          "credits": {
            "type": [
              "string",
              "null"
            ]
          },
          "file": {
            "type": "string",
            "format": "binary"
          },
          "kind": {
            "type": "string",
            "description": "Semantic role of the image. See [`ArtistImageKind`](api_types::artists::ArtistImageKind)."
          }
        }
      },
...
      "SongImageInfo": {
        "type": "object",
        "description": "An image attached to a song with its semantic role.",
        "required": [
          "id",
          "public_url",
          "kind"
        ],
        "properties": {
          "credits": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "kind": {
            "type": "string"
          },
          "public_url": {
            "type": "string",
            "description": "Publicly served URL for clients."
          }
        }
      },
...
      "SongResponse": {
        "type": "object",
        "description": "Full song metadata returned by detail endpoints.\n\nExcludes lyrics, fetch those via `GET /api/songs/{id}/lyrics` on demand.",
        "required": [
          "id",
          "title",
          "artists",
          "tags",
          "images"
        ],
        "properties": {
          "artists": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ArtistInfo"
            }
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "images": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SongImageInfo"
            }
          },
          "tags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TagInfo"
            }
          },
          "title": {
            "type": "string"
          }
        }
      },
...
      "SongTagAssignment": {
        "type": "object",
        "description": "A tag paired with its kind for application to a song.",
        "required": [
          "tag_id",
          "kind"
        ],
        "properties": {
          "kind": {
            "$ref": "#/components/schemas/SongTagKind"
          },
          "tag_id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
      "SongTagKind": {
        "type": "string",
        "description": "Valid kind values when applying a tag to a song.",
        "enum": [
          "genre",
          "source",
          "language",
          "misc"
        ]
      },
...
      "TagInfo": {
        "type": "object",
        "description": "A tag embedded in song responses.",
        "required": [
          "id",
          "name",
          "kind"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "kind": {
            "type": "string",
            "description": "Freeform category string (e.g. `\"genre\"`, `\"mood\"`)."
          },
          "name": {
            "type": "string"
          }
        }
      },
...
      "UpdateSongRequest": {
        "type": "object",
        "description": "Request body for `PUT /api/songs/{id}`.\n\nLyrics and images are excluded, manage them via their respective subresources.",
        "required": [
          "title",
          "artist_ids",
          "tags"
        ],
        "properties": {
          "artist_ids": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            }
          },
          "tags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SongTagAssignment"
            }
          },
          "title": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session"
      }
    }
  }
}